Playlist models is a dumpster fire
