+++ src/SpotifyApi.NetCore/Models/Album.cs
@@ -1,0 +1,47 @@
+using System;
+using Newtonsoft.Json;
+
+namespace SpotifyApi.NetCore
+{
+    public partial class Album
+    {
+        [JsonProperty("album_type")]
+        public string AlbumType { get; set; }
+
+        [JsonProperty("artists")]
+        public Artist[] Artists { get; set; }
+
+        [JsonProperty("available_markets")]
+        public string[] AvailableMarkets { get; set; }
+
+        [JsonProperty("external_urls")]
+        public ExternalUrls ExternalUrls { get; set; }
+
+        [JsonProperty("href")]
+        public Uri Href { get; set; }
+
+        [JsonProperty("id")]
+        public string Id { get; set; }
+
+        [JsonProperty("images")]
+        public Image[] Images { get; set; }
+
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        [JsonProperty("release_date")]
+        public DateTimeOffset ReleaseDate { get; set; }
+
+        [JsonProperty("release_date_precision")]
+        public string ReleaseDatePrecision { get; set; }
+
+        [JsonProperty("total_tracks")]
+        public long TotalTracks { get; set; }
+
+        [JsonProperty("type")]
+        public string Type { get; set; }
+
+        [JsonProperty("uri")]
+        public string Uri { get; set; }
+    }
+}
