+++ src/SpotifyApi.NetCore/Models/Common.cs
@@ -2,12 +2,24 @@ using Newtonsoft.Json;
 
 namespace SpotifyApi.NetCore
 {
+    /// <summary>
+    /// External URL object
+    /// </summary>
+    /// <remarks>
+    /// https://developer.spotify.com/documentation/web-api/reference/object-model/#external-url-object
+    /// </remarks>
     public partial class ExternalUrls
     {
         [JsonProperty("spotify")]
         public string Spotify { get; set; }
     }
 
+    /// <summary>
+    /// Followers object
+    /// </summary>
+    /// <remarks>
+    /// https://developer.spotify.com/documentation/web-api/reference/object-model/#followers-object
+    /// </remarks>
     public partial class Followers
     {
         [JsonProperty("href")]
@@ -17,6 +29,12 @@ namespace SpotifyApi.NetCore
         public int Total { get; set; }
     }
 
+    /// <summary>
+    /// Image object
+    /// </summary>
+    /// <remarks>
+    /// https://developer.spotify.com/documentation/web-api/reference/object-model/#image-object
+    /// </remarks>
     public partial class Image
     {
         [JsonProperty("height")]
@@ -27,5 +45,41 @@ namespace SpotifyApi.NetCore
 
         [JsonProperty("width")]
         public int Width { get; set; }
+    }
+
+    /// <summary>
+    /// User object (public)
+    /// </summary>
+    /// <remarks>
+    /// https://developer.spotify.com/documentation/web-api/reference/object-model/#user-object-public
+    /// </remarks>
+    public partial class User
+    {
+        [JsonProperty("display_name", NullValueHandling = NullValueHandling.Ignore)]
+        public string DisplayName { get; set; }
+
+        [JsonProperty("external_urls")]
+        public ExternalUrls ExternalUrls { get; set; }
+
+        [JsonProperty("followers")]
+        public Followers Followers { get; set; }
+
+        [JsonProperty("href")]
+        public string Href { get; set; }
+
+        [JsonProperty("id")]
+        public string Id { get; set; }
+
+        [JsonProperty("images")]
+        public Image[] Images { get; set; }
+
+        [JsonProperty("type")]
+        public string Type { get; set; }
+
+        [JsonProperty("uri")]
+        public string Uri { get; set; }
+
+        [JsonProperty("name", NullValueHandling = NullValueHandling.Ignore)]
+        public string Name { get; set; }
     }
 }
