--- src/SpotifyApi.NetCore/Models/PlaylistTracksResult.cs
@@ -1,54 +1,0 @@
-// Generated by https://quicktype.io
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
-namespace SpotifyApi.NetCore
-{
-    public partial class PlaylistTracksResult
-    {
-        [JsonProperty("href")]
-        public string Href { get; set; }
-
-        [JsonProperty("items")]
-        public TrackItem[] Items { get; set; }
-
-        [JsonProperty("limit")]
-        public long Limit { get; set; }
-
-        [JsonProperty("next")]
-        public string Next { get; set; }
-
-        [JsonProperty("offset")]
-        public long Offset { get; set; }
-
-        [JsonProperty("previous")]
-        public object Previous { get; set; }
-
-        [JsonProperty("total")]
-        public long Total { get; set; }
-    }
-
-    public partial class AddedBy
-    {
-        [JsonProperty("external_urls")]
-        public ExternalUrls ExternalUrls { get; set; }
-
-        [JsonProperty("href")]
-        public string Href { get; set; }
-
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        [JsonProperty("type")]
-        public string Type { get; set; }
-
-        [JsonProperty("uri")]
-        public string Uri { get; set; }
-
-        [JsonProperty("name", NullValueHandling = NullValueHandling.Ignore)]
-        public string Name { get; set; }
-    }
-}
