--- src/SpotifyApi.NetCore/Models/PlaylistsResult.cs
@@ -1,84 +1,0 @@
-// Generated by https://quicktype.io
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
-namespace SpotifyApi.NetCore
-{
-    public partial class PlaylistsResult
-    {
-        [JsonProperty("href")]
-        public string Href { get; set; }
-
-        [JsonProperty("items")]
-        public PlaylistItem[] Items { get; set; }
-
-        [JsonProperty("limit")]
-        public int Limit { get; set; }
-
-        [JsonProperty("next")]
-        public string Next { get; set; }
-
-        [JsonProperty("offset")]
-        public int Offset { get; set; }
-
-        [JsonProperty("previous")]
-        public string Previous { get; set; }
-
-        [JsonProperty("total")]
-        public int Total { get; set; }
-    }
-
-    public partial class PlaylistItem
-    {
-        [JsonProperty("collaborative")]
-        public bool Collaborative { get; set; }
-
-        [JsonProperty("external_urls")]
-        public ExternalUrls ExternalUrls { get; set; }
-
-        [JsonProperty("href")]
-        public string Href { get; set; }
-
-        [JsonProperty("id")]
-        public string Id { get; set; }
-
-        [JsonProperty("images")]
-        public Image[] Images { get; set; }
-
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("owner")]
-        public Owner Owner { get; set; }
-
-        [JsonProperty("primary_color")]
-        public string PrimaryColor { get; set; }
-
-        [JsonProperty("public")]
-        public bool Public { get; set; }
-
-        [JsonProperty("snapshot_id")]
-        public string SnapshotId { get; set; }
-
-        [JsonProperty("tracks")]
-        public PlaylistItemTracks Tracks { get; set; }
-
-        [JsonProperty("type")]
-        public string Type { get; set; }
-
-        [JsonProperty("uri")]
-        public string Uri { get; set; }
-    }
-
-    public partial class PlaylistItemTracks
-    {
-        [JsonProperty("href")]
-        public string Href { get; set; }
-
-        [JsonProperty("total")]
-        public int Total { get; set; }
-    }
-}
